+++ libsec/port/smallprimetest.c
@@ -2,7 +2,7 @@
 #include <mp.h>
 #include <libsec.h>
 
-static u32 smallprimes[] = {
+static ulong smallprimes[] = {
 	2,	3,	5,	7,	11,	13,	17,	19,	23,	29,
 	31,	37,	41,	43,	47,	53,	59,	61,	67,	71,
 	73,	79,	83,	89,	97,	101,	103,	107,	109,	113,
@@ -1007,7 +1007,7 @@ static u32 smallprimes[] = {
 
 //  return 1 if p is divisable by sp, 0 otherwise
 static int
-divides(mpint *dividend, u32 divisor)
+divides(mpint *dividend, ulong divisor)
 {
 	mpdigit d[2], q;
 	int i;
@@ -1026,7 +1026,7 @@ int
 smallprimetest(mpint *p)
 {
 	int i;
-	u32 sp;
+	ulong sp;
 
 	for(i = 0; i < nelem(smallprimes); i++){
 		sp = smallprimes[i];
